requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyYAML==6.0.3
Werkzeug==3.1.3
